3.8 - Convertendo tipos de variáveis

+++ adivinhacao.c
@@ -44,7 +44,7 @@ int main()
 
         tentativas++;
 
-        double pontosperdidos = (chute - numerosecreto) / 2.0;
+        double pontosperdidos = (chute - numerosecreto) / (double)2;
         pontos = pontos - pontosperdidos;
     }
